+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Move `ovlpr replace` to `rgr replace`
 * Move `linkr merge` to `rgr merge`
 
+* Rename .ranges to .rg
+
 ## 0.6.7 - 2022-04-24
 
 ## 0.6.6 - 2022-04-24